refactor a LivenessWalker, used by CoalesceLocals. This is more modular and lets CoalesceLocals only add the interference analysis, letting liveness analysis be reusable elsewhere
